added sorting events by date

+++ classes/Site/Controllers/Event.php
@@ -15,7 +15,7 @@ class Event {
 	}
 
     public function list() {
-        $result = $this->eventsTable->findAll();
+        $result = $this->eventsTable->findAllFutureDates('eventDate');
 
         $events = [];
           foreach ($result as $event) {
@@ -31,9 +31,12 @@ class Event {
                     'authorId' => $author['id']
 
                 ];
-      
+
+                
+
           }
-      
+
+         
         $title = 'Event list';
 
         $author = $this->authentication->getUser();
@@ -47,7 +50,7 @@ class Event {
 				];
         
     }
-
+    //checks auth and adds
     public function add() {
         $author = $this->authentication->getUser();
 
@@ -61,7 +64,7 @@ class Event {
 
         header('location: /event/list');
     }
-
+    //brings up form
     public function addpage() {
 
             $title = 'Add a new event';
@@ -71,17 +74,35 @@ class Event {
     }
 
     public function delete() {
+
+        $author = $this->authentication->getUser();
+
+        $event = $this->eventsTable->findById($_POST['eventId']);
+
+        if ($event['authorId'] != $author['id']) {
+            return;
+        }
+        
         $this->eventsTable->delete($_POST['eventId']);
     
         header('location: /event/list');
     }
 
     public function saveEdit() {
-      
+        $author = $this->authentication->getUser();
+
+        //added security from Ninja pg 493 PDF 363
+        if (isset($_GET['id'])) {
+            $event = $this->eventsTable->findById($_GET['id']);
+
+            if ($event['authorId'] != $author['id']) {
+                return;
+            }
+        }
             
         $event = $_POST['event'];
         //the above is from form, below is others
-        $event['authorId'] = 2;
+        $event['authorId'] = $author['id'];
 
         $this->eventsTable->save($event);
 
@@ -91,16 +112,19 @@ class Event {
 
     public function displayEdit() {
 
-            $event = $this->eventsTable->findById($_GET['id']);
+        $author = $this->authentication->getUser();
 
-            $title = 'Edit event';
+        $event = $this->eventsTable->findById($_GET['id']);
 
-            return ['template' => 'editevent.html.php', 
-                    'title' => $title,
-                    'variables' => [
-                        'event' => $event
-                        ]
-                    ];
+        $title = 'Edit event';
+
+        return ['template' => 'editevent.html.php', 
+                'title' => $title,
+                'variables' => [
+                    'event' => $event,
+                    'userId' => $author['id'] ?? null
+                    ]
+                ];
     }
 
 }
